started to move WebEntityAPIHelper from RenderableWebEntityItem to OffscreenQmlSurface
